+++ src/components/index.ts
@@ -1,16 +1,16 @@
-export * from './Button';
-export * from './Input';
-export * from './Loader';
-export * from './Modal';
-export * from './PasswordInput';
-export * from './Header';
-export * from './Footer';
-export * from './Nav';
-export * from './NavItem';
-export * from './UserList';
-export * from './DashboardAdmin';
-export * from './DashboardUser';
-export * from './UserCard';
+export * from './globals/Button';
+export * from './globals/Input';
+export * from './globals/Loader';
+export * from './globals/Modal';
+export * from './globals/PasswordInput';
+export * from './globals/UserCard';
+export * from './layout/Header';
+export * from './layout/Footer';
+export * from './layout/Nav';
+export * from './globals/NavItem';
+export * from './layout/UserList';
+export * from './layout/DashboardAdmin';
+export * from './layout/DashboardUser';
 export * from './Confirm';
 export * from './EditUserModal';
 export * from './Notification';
