+++ src/components/layout/DashboardUser/index.tsx
@@ -1,10 +1,10 @@
 import { useEffect, useState } from 'react';
-import { Loader } from '..';
+import { Loader } from '../..';
 import styles from './style.module.css';
-import { userServices } from '../../services';
+import { userServices } from '../../../services';
 import { useNavigate } from 'react-router-dom';
-import { User } from '../../interfaces/User';
-import { isTokenExpired } from '../../validations';
+import { User } from '../../../interfaces/User';
+import { isTokenExpired } from '../../../validations';
 
 export function DashboardUser() {
   const navigate = useNavigate();
